Revert "Support converting XML downloaded by some other tool"
This reverts commit a8d534ea856ed8dfb1c4519691a9b6b83e99fdbe.
