+++ download_posts.py
@@ -4,13 +4,11 @@ import os
 import requests
 from sys import exit as sysexit
 import xml.etree.ElementTree as ET
-from datetime import datetime
+from datetime import datetime, timedelta
 from dateutil.relativedelta import relativedelta
-from pathlib import Path
 
 from authentication import authenticated_request_params
 from utilities import save_json_file, save_text_file
-from utilities import xml_post_to_json as xml_to_json
 
 DATE_FORMAT = '%Y-%m'
 
@@ -55,6 +53,23 @@ def fetch_month_posts(year, month):
     return response.text
 
 
+def xml_to_json(xml):
+    def f(field):
+        return xml.findtext(field)
+
+    return {
+        'id': f('itemid'),
+        'date': f('logtime'),
+        'subject': f('subject') or '',
+        'body': f('event'),
+        'eventtime': f('eventtime'),
+        'security': f('security'),
+        'allowmask': f('allowmask'),
+        'current_music': f('current_music'),
+        'current_mood': f('current_mood')
+    }
+
+
 def download_posts():
     os.makedirs('posts-xml', exist_ok=True)
     os.makedirs('posts-json', exist_ok=True)
@@ -79,29 +94,6 @@ def download_posts():
     save_json_file('posts-json/all.json', json_posts)
 
     return json_posts
-
-
-def convert_posts():
-    xml_posts = Path('posts-xml')
-    all_json = Path('posts-json/all.json')
-    if not xml_posts.is_dir():
-        raise NotADirectoryError("Haven't got a posts-xml directory")
-    if not all_json.parent.exists():
-        all_json.parent.mkdir()
-    if not all_json.parent.is_dir():
-        raise NotADirectoryError(f'Not a directory: {all_json.parent}')
-    json_posts = []
-    maxid = 0
-    for xml in xml_posts.iterdir():
-        if xml.suffix == '.xml':
-            for json_post in map(xml_to_json, ET.fromstring(xml.read_text()).iter('entry')):
-                if json_post['id'] is None:
-                    maxid = json_post['id'] = maxid + 1
-                else:
-                    maxid = max((int(json_post['id']), maxid))
-                json_posts.append(json_post)
-
-    save_json_file(str(all_json), json_posts)
 
 
 if __name__ == '__main__':
